requirements.txt
weasyprint==45
Django==1.11.6
djangorestframework==3.6.4